New functions added
Going up or down floors now saves a temp copy of the previous roomset to retain the state when you return back to that roomset.
Implemented use of potions in and our of battle.
Implemented inital room description with look command to view it again if desired.
Implemented read command and signs that may be read.
Added blob and boss enemies.
Modified inital player stats.
Modified the pickle name for storing the current room.